+++ archiveIO/__init__.py
@@ -7,94 +7,35 @@ import tempfile
 from decorator import decorator
 
 
-@decorator
-def save(function, *args, **kwargs):
-    """
-    Decorator to support saving to compressed files for functions
-    whose first argument is the targetPath
-
-    If the first argument ends with a recognized extension,
-    the decorator runs the function in a temporary folder and
-    compresses the resulting output to targetPath.
-
-    Archive format is determined by file extension:
-    .zip .tar.gz .tar.bz2 .tar
-    """
-    targetPath = kwargs.get('targetPath', args[0])
-    try:
-        archive = Archive(targetPath)
-    # If we did not recognize the extension, run function as usual
-    except ArchiveError:
-        return function(*args, **kwargs)
-    # Make temporaryFolder
-    with TemporaryFolder() as temporaryFolder:
-        # Run function in temporaryFolder
-        temporaryPath = os.path.join(temporaryFolder, archive.getName())
-        function(temporaryPath, *args[1:], **kwargs)
-        filePaths = []
-        # Walk function output
-        for rootPath, directories, fileNames in os.walk(temporaryFolder):
-            # For each file,
-            for fileName in fileNames:
-                filePaths.append(os.path.join(rootPath, fileName))
-        # Save
-        return archive.save(filePaths, temporaryFolder)
-
-
-@decorator
-def load(function, *args, **kwargs):
-    """
-    Decorator to support loading from compressed files for functions
-    whose first argument is the sourcePath
-
-    If the first argument ends with a recognized extension,
-    the decorator uncompresses sourcePath to a temporary folder
-    and runs the function on each resulting file until it succeeds.
-
-    Archive format is determined by file extension:
-    .zip .tar.gz .tar.bz2 .tar
-    """
-    sourcePath = kwargs.get('sourcePath', args[0])
-    try:
-        archive = Archive(sourcePath)
-    # If we did not recognize the extension, run function as usual
-    except ArchiveError:
-        return function(*args, **kwargs)
-    # Make temporaryFolder
-    with TemporaryFolder() as temporaryFolder:
-        # For each uncompressed filePath,
-        for filePath in archive.load(temporaryFolder):
-            # Run function and exit if successful
-            errors = []
-            try:
-                return function(filePath, *args[1:], **kwargs)
-            except Exception, error:
-                errors.append(str(error))
-        else:
-            raise IOError('Could not run %s on any file in %s:\n%s' % (function, sourcePath, '\n'.join(errors)))
-
-
 class Archive(object):
-    'Archive processor'
     
-    def __init__(self, path):
-        'Prepare'
-        pathLower = path.lower()
-        # Try to recognize the extension,
-        for extension, make_consumer, make_generator in extensionPacks:
-            # If we have a matching extension, exit loop
-            if pathLower.endswith(extension):
-                break
-        # If we did not recognize the extension, raise exception
+    def __init__(self, path, extension=None):
+        """
+        The path can be a string or file-like object.
+        The extension determines the archive format.
+        """
+        # If the extension is specified, use it
+        if extension:
+            try:
+                make_consumer, make_generator = dict(EXTENSION_PACKS)[extension.lower()]
+            except KeyError:
+                raise ArchiveError('Could not recognize file extension: %s' % extension)
+        # If path is a file-like object, raise Exception
+        elif hasattr(path, 'read'):
+            raise ArchiveError('Must specify file extension when using a file-like object')
+        # If path is a string, try to recognize the extension
         else:
-            raise ArchiveError('Could not recognize compression format from file extension')
-        # Remove matching extension from filename
-        name = os.path.basename(path[:pathLower.rfind(extension)])
+            pathLower = path.lower()
+            for extension, (make_consumer, make_generator) in EXTENSION_PACKS:
+                if pathLower.endswith(extension):
+                    break
+            else:
+                raise ArchiveError('Could not recognize archive format from file extension')
         # Set
         self.__path = path
+        self.__extension = extension
         self.__make_consumer = make_consumer
         self.__make_generator = make_generator
-        self.__name = name
 
     def save(self, filePaths, basePath=''):
         """
@@ -104,26 +45,27 @@ class Archive(object):
         # Convert filePaths into a list if it isn't one already
         if not hasattr(filePaths, '__iter__'):
             filePaths = [filePaths]
-        # Enter consumer
+        baseIndex = len(os.path.relpath(basePath)) + 1 if basePath else 0
         consumer = self.__make_consumer(self.__path)
         consumer.next()
-        # For each filePath,
-        for filePath in filePaths:
+        for filePath in expand_paths(filePaths):
             # Truncate filePath into relativePath
-            relativeIndex = len(basePath) + 1 if basePath else 0
-            consumer.send((filePath, filePath[relativeIndex:]))
-        # Exit consumer
+            consumer.send((filePath, os.path.relpath(filePath)[baseIndex:]))
         consumer.close()
-        # Return
+        # If path is a file-like object, prepare it
+        if hasattr(self.__path, 'read'):
+            self.__path.reset()
         return self.__path
 
     def load(self, targetFolder):
         'Uncompress archivePath to a targetFolder.'
+        # If path is a file-like object, prepare it
+        if hasattr(self.__path, 'read'):
+            self.__path.reset()
         return self.__make_generator(self.__path, targetFolder)
 
-    def getName(self):
-        'Return archive filename after stripping file extension'
-        return self.__name
+    def get_extension(self):
+        return self.__extension
 
 
 class ArchiveError(Exception):
@@ -147,6 +89,101 @@ class TemporaryFolder(object):
         shutil.rmtree(self.temporaryFolder)
 
 
+@decorator
+def save(function, *args, **kw):
+    """
+    Decorator to support saving to compressed files for functions
+    whose first argument is the targetPath
+
+    If the first argument ends with a recognized extension,
+    the decorator runs the function in a temporary folder and
+    compresses the resulting output to targetPath.
+
+    Archive format is determined by file extension:
+    .zip .tar.gz .tar.bz2 .tar
+    """
+    targetPath = kw.get('targetPath', args[0])
+    try:
+        archive = Archive(targetPath, extension=kw.get('targetExtension'))
+    # If we did not recognize the extension, run function as usual
+    except ArchiveError:
+        return function(*args, **kw)
+    targetName = kw.get('targetName')
+    if not targetName:
+        # If path is a file-like object, raise exception
+        if hasattr(targetPath, 'read'):
+            raise ArchiveError('Must specify targetName when using a file-like object')
+        # If path is a string, remove matching extension from filename
+        else:
+            targetName = os.path.basename(targetPath[:targetPath.lower().rfind(archive.get_extension())])
+    # Make temporaryFolder
+    with TemporaryFolder() as temporaryFolder:
+        # Run function in temporaryFolder
+        function(os.path.join(temporaryFolder, targetName), *args[1:], **kw)
+        # Save
+        return archive.save(walk_paths(temporaryFolder), temporaryFolder)
+
+
+@decorator
+def load(function, *args, **kw):
+    """
+    Decorator to support loading from compressed files for functions
+    whose first argument is the sourcePath
+
+    If the first argument ends with a recognized extension,
+    the decorator uncompresses sourcePath to a temporary folder
+    and runs the function on each resulting file until it succeeds.
+
+    Archive format is determined by file extension:
+    .zip .tar.gz .tar.bz2 .tar
+    """
+    sourcePath = kw.get('sourcePath', args[0])
+    try:
+        archive = Archive(sourcePath)
+    # If we did not recognize the extension, run function as usual
+    except ArchiveError:
+        return function(*args, **kw)
+    # Make temporaryFolder
+    with TemporaryFolder() as temporaryFolder:
+        errors = []
+        # For each uncompressed filePath,
+        for filePath in archive.load(temporaryFolder):
+            # Run function and exit if successful
+            try:
+                return function(filePath, *args[1:], **kw)
+            except Exception, error:
+                errors.append(str(error))
+        else:
+            raise IOError('Could not run %s on any file in %s:\n%s' % (function, sourcePath, '\n'.join(errors)))
+
+
+def expand_paths(paths):
+    'Expand folderPaths'
+    filePaths = []
+    for path in paths:
+        if os.path.isdir(path):
+            filePaths.extend(walk_paths(path))
+        else:
+            filePaths.append(path)
+    return set(filePaths)
+
+
+def walk_paths(path):
+    'Yield filePaths one by one from the specified path'
+    for rootPath, folderNames, fileNames in os.walk(path):
+        for fileName in fileNames:
+            yield os.path.join(rootPath, fileName)
+
+
+def open_tarfile(targetPath, mode):
+    valueByKey = dict(mode=mode)
+    if hasattr(targetPath, 'read'):
+        valueByKey['fileobj'] = targetPath
+    else:
+        valueByKey['name'] = targetPath
+    return tarfile.open(**valueByKey)
+
+
 def make_consumer_tar(targetPath, mode='w'):
     'Save .tar file'
     def filter_(tarInfo):
@@ -154,9 +191,9 @@ def make_consumer_tar(targetPath, mode='w'):
         tarInfo.uid = tarInfo.gid = 0
         tarInfo.uname = tarInfo.gname = 'root'
         return tarInfo
-    with tarfile.open(targetPath, mode) as targetFile:
+    with open_tarfile(targetPath, mode) as targetFile:
         while True:
-            filePath, relativePath = (yield)
+            filePath, relativePath = yield
             targetFile.add(filePath, relativePath, recursive=False, filter=filter_)
 
 
@@ -174,13 +211,13 @@ def make_consumer_zip(targetPath):
     'Save .zip file'
     with zipfile.ZipFile(targetPath, 'w', zipfile.ZIP_DEFLATED) as targetFile:
         while True:
-            filePath, relativePath = (yield)
+            filePath, relativePath = yield
             targetFile.write(filePath, relativePath)
 
 
 def make_generator_tar(sourcePath, temporaryFolder):
     'Load .tar.gz or .tar.bz2 or .tar file'
-    with tarfile.open(sourcePath, 'r') as sourceFile:
+    with open_tarfile(sourcePath, 'r') as sourceFile:
         sourceFile.extractall(temporaryFolder)
         for relativePath in sourceFile.getnames():
             yield os.path.join(temporaryFolder, relativePath)
@@ -194,9 +231,9 @@ def make_generator_zip(sourcePath, temporaryFolder):
             yield os.path.join(temporaryFolder, relativePath)
 
 
-extensionPacks = [
-    ('.zip', make_consumer_zip, make_generator_zip),
-    ('.tar.gz', make_consumer_tar_gz, make_generator_tar),
-    ('.tar.bz2', make_consumer_tar_bz2, make_generator_tar),
-    ('.tar', make_consumer_tar, make_generator_tar),
+EXTENSION_PACKS = [
+    ('.zip', (make_consumer_zip, make_generator_zip)),
+    ('.tar.gz', (make_consumer_tar_gz, make_generator_tar)),
+    ('.tar.bz2', (make_consumer_tar_bz2, make_generator_tar)),
+    ('.tar', (make_consumer_tar, make_generator_tar)),
 ]